Fix specs after changing implementation
